requirements.txt
# Core ML & AI
tensorflow==2.16.1
tf-keras==2.16.0
transformers==4.41.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
librosa==0.10.1
scikit-learn==1.4.2
opencv-python==4.9.0.80

# Spotify Integration
spotipy==2.25.0

# Audio Processing
sounddevice==0.4.7
soundfile==0.12.1
pydub==0.25.1

# Text Sentiment
vaderSentiment==3.3.2

# Web App & Utilities
streamlit==1.35.0
python-dotenv==1.0.1
numpy<2.0 # Pin numpy to avoid v2.0 compatibility issues

# --- Packages below are dependencies but good to pin ---
# h5py is for saving Keras models
h5py==3.11.0
# For Hugging Face models
huggingface-hub==0.23.0
safetensors==0.4.3
tokenizers==0.19.1